Always send RUM data to Spoor when using Toggler flags
RUM data is only sent to spoor if you're randomly allocated to the
sampling group (based on your SpoorID).

However this means it's not easy for engineers to see (in dev tools)
what RUM data is being sent. They have to be randomly allocated to the
sample group.

With this change, any time the Toggle flags cookie "next-flags" is set,
we send the data. That makes it easy to see RUM data while developing
any feature.