Switch from Perfume.js to Google Web Vitals
This switches us over to Google Web Vitals. I decided to rewrite this
instead of using #87 for a couple of reasons:

  1. There are some broader format changes in that PR which made it
     difficult to reason about the differences between the two libs.
     `src/client/trackers/realUserMonitoringForPerformance.js` has
     28 rather than 99 changes.

  2. There's been a new major version of the `web-vitals` package since
     the old PR was written.

  3. I wanted to try out the experimental metrics so that we had a more
     complete comparison to Perfume: `onFCP` and `onTTFB`.

I still need to test in a browser, and I'll be doing this by installing
this pull request and testing in a local running app to make sure that
metrics still make it through in the expected format.

Relates-To: https://financialtimes.atlassian.net/browse/FTDCS-294