Request all metric changes from web-vitals
This came about as a result of a lot of testing with the home page
running locally. The oTracking event was never firing because `onCLS`
was just never being called.

In order to match with Perfume, we need to tell web-vitals to always
report Cumulative Layout Shift, even if it's set to zero.

This seems a bit odd but it matches the behaviour of Perfume based on my
local testing. I'm happy to see if we get the same in prod.